Versão de apresentação do netb

+++ NetB/Models/DTO/TarefasDTO.cs
@@ -19,8 +19,8 @@ namespace NetB.Models.DTO
         public string observacoes { get; set; }
         public int dias_estimados { get; set; }
         public int dias_trabalhados { get; set; }
-        public int valor_estimado { get; set; }
-        public int valor_utilizado { get; set; }
+        public double valor_estimado { get; set; }
+        public double valor_utilizado { get; set; }
         public string statusCor { get; set; }
     }
 }
